+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/TopicsAPIController.java
@@ -12,11 +12,13 @@ import de.digitalcollections.model.identifiable.entity.Topic;
 import de.digitalcollections.model.identifiable.resource.FileResource;
 import de.digitalcollections.model.list.paging.PageResponse;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import java.util.List;
 import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -36,6 +38,27 @@ public class TopicsAPIController extends AbstractEntitiesController<Topic, Cudam
     super(client.forTopics(), client, languageService);
   }
 
+  @PostMapping("/api/topics/{uuid:" + ParameterHelper.UUID_PATTERN + "}/entities")
+  public ResponseEntity addEntities(@PathVariable UUID uuid, @RequestBody List<Entity> entities)
+      throws TechnicalException {
+    boolean successful = ((CudamiTopicsClient) service).addEntities(uuid, entities);
+    if (successful) {
+      return new ResponseEntity<>(successful, HttpStatus.OK);
+    }
+    return new ResponseEntity<>(successful, HttpStatus.NOT_FOUND);
+  }
+
+  @PostMapping("/api/topics/{uuid:" + ParameterHelper.UUID_PATTERN + "}/fileresources")
+  public ResponseEntity addFileResources(
+      @PathVariable UUID uuid, @RequestBody List<FileResource> fileResources)
+      throws TechnicalException {
+    boolean successful = ((CudamiTopicsClient) service).addFileResources(uuid, fileResources);
+    if (successful) {
+      return new ResponseEntity<>(successful, HttpStatus.OK);
+    }
+    return new ResponseEntity<>(successful, HttpStatus.NOT_FOUND);
+  }
+
   @GetMapping("/api/topics/new")
   @ResponseBody
   public Topic create() throws TechnicalException {
@@ -44,7 +67,7 @@ public class TopicsAPIController extends AbstractEntitiesController<Topic, Cudam
 
   @GetMapping("/api/topics/{uuid:" + ParameterHelper.UUID_PATTERN + "}/entities")
   @ResponseBody
-  public BTResponse<Entity> findRelatedEntities(
+  public BTResponse<Entity> findEntities(
       @PathVariable UUID uuid,
       @RequestParam(name = "offset", required = false, defaultValue = "0") int offset,
       @RequestParam(name = "limit", required = false, defaultValue = "10") int limit,
@@ -70,7 +93,7 @@ public class TopicsAPIController extends AbstractEntitiesController<Topic, Cudam
 
   @GetMapping("/api/topics/{uuid:" + ParameterHelper.UUID_PATTERN + "}/fileresources")
   @ResponseBody
-  public BTResponse<FileResource> findRelatedFileResources(
+  public BTResponse<FileResource> findFileResources(
       @PathVariable UUID uuid,
       @RequestParam(name = "offset", required = false, defaultValue = "0") int offset,
       @RequestParam(name = "limit", required = false, defaultValue = "10") int limit,
@@ -96,7 +119,7 @@ public class TopicsAPIController extends AbstractEntitiesController<Topic, Cudam
 
   @GetMapping("/api/topics/{uuid:" + ParameterHelper.UUID_PATTERN + "}/topics")
   @ResponseBody
-  public BTResponse<Topic> findSubtopic(
+  public BTResponse<Topic> findSubtopics(
       @PathVariable UUID uuid,
       @RequestParam(name = "offset", required = false, defaultValue = "0") int offset,
       @RequestParam(name = "limit", required = false, defaultValue = "10") int limit,
@@ -116,7 +139,7 @@ public class TopicsAPIController extends AbstractEntitiesController<Topic, Cudam
   @SuppressFBWarnings
   @GetMapping("/api/topics")
   @ResponseBody
-  public BTResponse<Topic> findTop(
+  public BTResponse<Topic> findTopTopics(
       @RequestParam(name = "offset", required = false, defaultValue = "0") int offset,
       @RequestParam(name = "limit", required = false, defaultValue = "10") int limit,
       @RequestParam(name = "search", required = false) String searchTerm,
@@ -135,6 +158,35 @@ public class TopicsAPIController extends AbstractEntitiesController<Topic, Cudam
   @ResponseBody
   public Topic getByUuid(@PathVariable UUID uuid) throws TechnicalException {
     return service.getByUuid(uuid);
+  }
+
+  /*
+   * Used in templates/topics/view.html
+   */
+  @DeleteMapping("/api/topics/{topicUuid}/entities/{entityUuid}")
+  @ResponseBody
+  public ResponseEntity removeEntity(@PathVariable UUID topicUuid, @PathVariable UUID entityUuid)
+      throws TechnicalException {
+    boolean successful = ((CudamiTopicsClient) service).removeEntity(topicUuid, entityUuid);
+    if (successful) {
+      return new ResponseEntity<>(successful, HttpStatus.OK);
+    }
+    return new ResponseEntity<>(successful, HttpStatus.NOT_FOUND);
+  }
+
+  /*
+   * Used in templates/topics/view.html
+   */
+  @DeleteMapping("/api/topics/{topicUuid}/fileresources/{fileResourceUuid}")
+  @ResponseBody
+  public ResponseEntity removeFileResource(
+      @PathVariable UUID topicUuid, @PathVariable UUID fileResourceUuid) throws TechnicalException {
+    boolean successful =
+        ((CudamiTopicsClient) service).removeFileResource(topicUuid, fileResourceUuid);
+    if (successful) {
+      return new ResponseEntity<>(successful, HttpStatus.OK);
+    }
+    return new ResponseEntity<>(successful, HttpStatus.NOT_FOUND);
   }
 
   @PostMapping("/api/topics")