+++ CHANGELOG.md
@@ -37,7 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Add persistence support for new field `Headword.labelNormalized`
 - Add sorting to Buckets and Bucket objects handling in `HeadwordRepositoryImpl` 
 - Add action links to object's view and edit under topic
-    
+- Add topic add/remove entities/fileresources
+
 ### Changed
 
 - WEMI parts: Manifestations, Works
@@ -70,6 +71,7 @@ by `"und"`
 - Fix data language handling in admin webapp
 - Fix visibility of long text in collection view
 - Fix website label rendering
+- Fix collection add digitalobjects
 
 ### Removed